apply pr review suggestions
